build: split main E2E test into smaller files
Ref. metriport/metriport-internal#1634
