+++ packages/api/src/__tests__/e2e/mapi/parts/consolidated/consolidated-template.html
@@ -183,7 +183,7 @@
       <td>
         <span class="header-label">Name</span>
         <span>
-          Junhdjjdkksuyujebeb Xamuscaeyttyworo
+          Junhdjjdkksuyujebeb <%= lastName %>
         </span>
       </td>
     </tr>
