+++ packages/api/src/__tests__/e2e/mapi/parts/consolidated/consolidated-template.json
@@ -1,0 +1,1 @@
+{"resourceType":"Bundle","total":3,"type":"searchset","entry":[{"resource":{"resourceType":"AllergyIntolerance","id":"<%= allergyId %>","meta":{"versionId":"1","lastUpdated":"<%= date %>","source":"#is1IVeJ3jZeDIOAr","profile":["http://hl7.org/fhir/us/core/StructureDefinition/us-core-allergyintolerance"]},"contained":[{"resourceType":"Patient","id":"<%= patientId %>","meta":{"source":"#aFlyTk4ikkVYq0UC"},"name":[{"family":"<%= lastName %>","given":["Junhdjjdkksuyujebeb"]}],"telecom":[{"system":"phone","value":"<%= phone %>"},{"system":"email","value":"<%= email %>"}],"gender":"male","birthDate":"1900-01-01","address":[{"line":["123 test ave"],"city":"Miami","state":"FL","postalCode":"12345","country":"USA"}]}],"clinicalStatus":{"coding":[{"system":"http://terminology.hl7.org/CodeSystem/allergyintolerance-clinical","code":"active"}]},"verificationStatus":{"coding":[{"system":"http://terminology.hl7.org/CodeSystem/allergyintolerance-verification","code":"confirmed"}]},"type":"allergy","category":["environment"],"criticality":"low","code":{"coding":[{"system":"http://snomed.info/sct","code":"419199007","display":"Allergy to substance (finding)"}],"text":"Allergy to substance (finding)"},"patient":{"reference":"Patient/<%= patientId %>"},"recordedDate":"2010-09-03T03:10:10-05:00","reaction":[{"substance":{"text":"Pollen"},"manifestation":[{"coding":[{"system":"http://snomed.info/sct","code":"271807003","display":"Eruption of skin (disorder)"}],"text":"Eruption of skin (disorder)"}]}]}},{"resource":{"resourceType":"DocumentReference","id":"<%= documentId %>","meta":{"versionId":"1","lastUpdated":"2024-06-08T18:22:20.913+00:00","source":"#is1IVeJ3jZeDIOAr"},"contained":[{"resourceType":"Organization","id":"2.16.840.1.113883.3.9621.5.2004","name":"Metriport 2004"},{"resourceType":"Patient","id":"<%= patientId %>","meta":{"source":"#aFlyTk4ikkVYq0UC"},"name":[{"family":"<%= lastName %>","given":["Junhdjjdkksuyujebeb"]}],"telecom":[{"system":"phone","value":"<%= phone %>"},{"system":"email","value":"<%= email %>"}],"gender":"male","birthDate":"1900-01-01","address":[{"line":["123 test ave"],"city":"Miami","state":"FL","postalCode":"12345","country":"USA"}]}],"extension":[{"url":"https://public.metriport.com/fhir/StructureDefinition/data-source.json","valueCoding":{"system":"https://public.metriport.com/fhir/StructureDefinition/data-source.json","code":"METRIPORT"}}],"masterIdentifier":{"system":"urn:ietf:rfc:3986","value":"2.16.840.1.113883.3.9621.5.2004.666001"},"identifier":[{"use":"official","system":"urn:ietf:rfc:3986","value":"2.16.840.1.113883.3.9621.5.2004.666001"}],"status":"current","type":{"coding":[{"system":"http://loinc.org/","code":"75622-1","display":"HIV 1 and 2 tests - Meaningful Use set"}]},"subject":{"reference":"Patient/<%= patientId %>","type":"Patient"},"author":[{"reference":"#2.16.840.1.113883.3.9621.5.2004","type":"Organization"}],"description":"Summarization Of Episode Notes - provided by Metriport","content":[{"attachment":{"contentType":"application/json","url":"http://somedomain.com/fhir/R4/Binary/<%= binaryId %>"}}],"context":{"event":[{"coding":[{"system":"http://snomed.info/sct","code":"62479008","display":"AIDS"}],"text":"AIDS"}],"period":{"start":"2022-10-05T22:00:00.000Z","end":"2022-10-05T23:00:00.000Z"},"sourcePatientInfo":{"id":"<%= patientId %>","type":"Patient"}}}},{"resource":{"resourceType":"Patient","id":"<%= patientId %>","meta":{"versionId":"1","lastUpdated":"2024-06-08T18:22:18.750+00:00","source":"#aFlyTk4ikkVYq0UC"},"name":[{"family":"<%= lastName %>","given":["Junhdjjdkksuyujebeb"]}],"telecom":[{"system":"phone","value":"<%= phone %>"},{"system":"email","value":"<%= email %>"}],"gender":"male","birthDate":"1900-01-01","address":[{"line":["123 test ave"],"city":"Miami","state":"FL","postalCode":"12345","country":"USA"}]}}]}
