+++ packages/api/src/__tests__/e2e/mapi/shared.ts
@@ -1,5 +1,13 @@
 import { MedplumClient } from "@medplum/core";
-import { MetriportMedicalApi } from "@metriport/api-sdk";
+import {
+  AllergyIntolerance,
+  Binary,
+  Bundle,
+  DocumentReference,
+  Resource,
+} from "@medplum/fhirtypes";
+import { Facility, MetriportMedicalApi, PatientDTO } from "@metriport/api-sdk";
+import { PatientWithId } from "@metriport/core/external/fhir/__tests__/patient";
 import { baseURL, testApiKey } from "../shared";
 
 export const ACCOUNT_PATH = "/internal/admin/cx-account";
@@ -10,3 +18,16 @@ export const fhirHeaders = { headers: { "x-api-key": testApiKey } };
 
 export const fhirApi = new MedplumClient({ baseUrl: baseURL });
 export const medicalApi = new MetriportMedicalApi(testApiKey, { baseAddress: baseURL });
+
+export type E2eContext = {
+  facility?: Facility | undefined;
+  patient?: PatientDTO | undefined;
+  patientFhir?: PatientWithId | undefined;
+  consolidatedPayload?: Bundle<Resource> | undefined;
+  allergyIntolerance?: AllergyIntolerance | undefined;
+  documentReference?: DocumentReference | undefined;
+  binary?: Binary | undefined;
+  url?: string | undefined;
+  mrContentBuffer?: Buffer | undefined;
+  expectedWebhookMeta?: Record<string, string> | undefined;
+};
